Update Gist link in README.md
Updated the link to the example `resume.json` hosted on Gist in the README.md file. The previous link mistakenly had `.js` at the end of the URL, this has been corrected in this commit. With this fix, users should be able to easily navigate to the example `resume.json` file.